// apps/web/src/api/generated/model/completeQuestionBlueprintDraftWorkbookEditorUploadResponse.ts
/**
 * Generated by orval v8.17.0 🍺
 * Do not edit manually.
 * Lemma API
 * OpenAPI spec version: 0.1.0
 */
import type { WorkbookEditorOutputFile } from "./workbookEditorOutputFile.ts";

export interface CompleteQuestionBlueprintDraftWorkbookEditorUploadResponse {
  editorOutputFile: WorkbookEditorOutputFile;
}