// apps/web/src/api/generated/model/createQuestionBlueprintDraftWorkbookEditorUploadResponse.ts
/**
 * Generated by orval v8.17.0 🍺
 * Do not edit manually.
 * Lemma API
 * OpenAPI spec version: 0.1.0
 */
import type { WorkbookEditorUpload } from "./workbookEditorUpload.ts";
import type { WorkbookEditorUploadUrl } from "./workbookEditorUploadUrl.ts";

export interface CreateQuestionBlueprintDraftWorkbookEditorUploadResponse {
  upload: WorkbookEditorUpload;
  uploadUrl: WorkbookEditorUploadUrl;
}