// apps/web/src/api/generated/model/listQuestionBlueprintDraftsResponse.ts
/**
 * Generated by orval v8.16.0 🍺
 * Do not edit manually.
 * Lemma API
 * OpenAPI spec version: 0.1.0
 */
import type { QuestionBlueprintDraft } from "./questionBlueprintDraft.ts";

export interface ListQuestionBlueprintDraftsResponse {
  drafts: QuestionBlueprintDraft[];
  /** @nullable */
  nextCursor: string | null;
}