// apps/web/src/api/generated/model/publishQuestionBlueprintDraftResponse.ts
/**
 * Generated by orval v8.17.0 🍺
 * Do not edit manually.
 * Lemma API
 * OpenAPI spec version: 0.1.0
 */
import type { QuestionBlueprint } from "./questionBlueprint.ts";
import type { QuestionBlueprintDraft } from "./questionBlueprintDraft.ts";
import type { QuestionBlueprintVersion } from "./questionBlueprintVersion.ts";

export interface PublishQuestionBlueprintDraftResponse {
  draft: QuestionBlueprintDraft;
  questionBlueprint: QuestionBlueprint;
  questionBlueprintVersion: QuestionBlueprintVersion;
}